From Ian MacIntosh:
Modifications that it doesn't use the uscsi commands anymore and so
all IDE and SCSI drives go through the same code path - a major
simplification.

This works with all the CD drives that work with Solaris (IDE or SCSI)
as the ioctl's for CDROMCDXA in Solaris just end up doing uscsi calls
anyway, but directly from the kernel. The added advantage of not using
uscsi is that the user doesn't need to be root - which is important in
Sunray environments.