Use the LTLIBICONV variable rather than LIBICONV.
With this change, instead of using the fully qualified path to the
shared object (or the one that the configure think is the fully
qualified path), the path where the library is found will be added to
the search path and just a generic -liconv will be used.

The old variable would be fooled up when /usr/lib/libiconv.so is an LD
script that redirects to /lib/libiconv.so, causing failures with some
linkers.

Also, replace @LIBICONV@ for libcdio itself also with $(LTLIBICONV) or
it will fail to link against on uClibc.