cdda-player configure patch from Stef #54191
From Stef:
I came up with a patch that I believe will work for everyone.
It uses AC_SEARCH_LIBS, instead of AC_CHECK_LIB. So, if keypad() is
defined in -lncurses or -lcurses, it does exactly what it does
now. However, if -ltinfo is needed, it is added to CDDA_PLAYER_LIBS
along with the curses library.