Add Rock Ridge deep directory support
In order to locate the target of a child link, we need to be able to call
iso9660_[i]fs_find_lsn() from within get_rock_ridge_filename() which means
that we need to add an additional image parameter for the handful of calls
that are used during directory traversing. We also need to use a special
version of the call, that works with DD disabled, since we need to process
RE entries, i.e the directories that contain the CL target, as well as
ignore CL elements altogether, so that we don't end up with an endless
recursion.

All of the above can be achieved by also carrying a 'DISABLE_RR_DD'
flag in the image. However since we need to ensure that we handle both
iso9660_t as well as CdIo_t as the image, we modify both structures so
that they have a common header which we can reference.

Note that the approach of turning Rock Ridge extensions altogether during
LSN seek (through the use of the iso9660_t extension mask for instance)
is not appropriate as we do want to be able to retrieve other Rock Ridge
attributes, such as long filenames, for the child link target, hence the
requirement to being able to disable deep directory alone.