Add Rock Ridge deep directory test
The test is designed to test more than one level of deep directories.

Also the ISO is designed so that the folder with deep directory children
(/1/) appears before the target (/rr_moved/), so that any endless
recursion during lookup of the CL target will be tested.
In other words, this validates that the deep directory lookup doesn't
only work if the CL target can be listed before the CL source.