Use getmntent/setmntent for reading mounts
Since fields in the mtab and fstab files are separated by
whitespace, octal escapes are used to represent the characters
space (\040), tab (\011), newline (\012), and backslash (\\) in
those files when they occur in one of the four strings in a
mntent structure.  The routines addmntent() and getmntent() will
convert from string representation to escaped representation and
back.  When converting from escaped representation, the sequence
\134 is also converted to a backslash.