Add processing of Rock Ridge CE records
Use the processed cont_extent to read the relevant continuation block
once the current extensions have been processed. The code is designed
to handle chained CE blocks if needed, and we also add code to detect
ISOs that are abusing the specs to loop CEs.
Also clean up the CHECK...() macros and fix some whitespace issues.