Use non widestring API calls for MSVC where required
When compiling for MSVC with Unicode enabled, the compiler tries to use widestring API calls
unless otherwise specified. So make sure we enforce the use of non widestring (A suffixed)
calls when we pass char* parameters.

Also fix a warning related to the declaration of is_cdrom_aspi() and silence a warning about
GetVersion() being deprecated.